\name{icsurvROC-package}
\alias{icsurvROC-package}
\docType{package}
\title{Interval Censored Survival ROC}
\description{Nonparametric and semiparametric estimations of the time-dependent ROC curve for the interval-censored failure time data.}
\details{
\tabular{ll}{
Package: \tab icsurvROC\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-11-24\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [cre]

Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yunro Chung, Tianxi Cai, Yingye Zheng, Estimating Diagnostic Accuracy Measures for Current Status Survival Data with Application to Prostate Cancer Active Surveillance Study (in progress)}
\keyword{Semiparametric Cox proportional hazards, local NPMLE, maximum likelihood croos validation}
